finishing up basic start to app
this first version includes the base, where the user is able
to instantiate a bingo instance, load a file, and then
add other local files to it. The current board has scoring
for rows, columns, and diagonals, and also supports clearing
and resetting, or choosing a new list. It would be nice to have
save to file of the board, caching, and a color chooser.

Signed-off-by: vsoch <vsochat@stanford.edu>